+++ crawler.h
@@ -34,7 +34,7 @@ int stop_item_crawler_thread(bool wait);
 int init_lru_crawler(void *arg);
 enum crawler_result_type lru_crawler_crawl(char *slabs, enum crawler_run_type,
         void *c, const int sfd, unsigned int remaining);
-int lru_crawler_start(uint8_t *ids, uint32_t remaining,
+int lru_crawler_start(uint16_t *ids, uint32_t remaining,
                              const enum crawler_run_type type, void *data,
                              void *c, const int sfd);
 void lru_crawler_pause(void);
